Create text to image converter
The Text to Image Converter is a simple Java application that allows users to convert text into an image format. This can be useful for creating graphical representations of text for use in applications, social media, or graphic design projects. The program uses Java's AWT and Swing libraries to create an image containing the specified text and allows users to save it as a PNG file.